fix bug exit fec

+++ app/src/main/java/com/example/lifefit/UserLoginActivity.java
@@ -32,6 +32,7 @@ public class UserLoginActivity extends AppCompatActivity implements View.OnClick
     private EditText et_password;
     private Button b_login;
     private TextView link_signup;
+    private long pressedTime;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -124,10 +125,19 @@ public class UserLoginActivity extends AppCompatActivity implements View.OnClick
         }
     }
 
-    /**@Override
+    @Override
     public void onBackPressed() {
-        android.os.Process.killProcess(android.os.Process.myPid());
-        finish();
-    }**/
+
+        if (pressedTime + 2000 > System.currentTimeMillis()) {
+            //super.onBackPressed();
+            finishAffinity();
+            android.os.Process.killProcess(android.os.Process.myPid());
+
+
+        } else {
+            Toast.makeText(getBaseContext(), "Press back again to exit", Toast.LENGTH_SHORT).show();
+        }
+        pressedTime = System.currentTimeMillis();
+    }
 
 }
